Added some better diagnostics.
Debug video now includes the tracks that we found but threw away. Also
added some (temporary) stuff to print out when & why we threw out a
track. Basically the "missing face detections" from last commit were
just tracks getting thrown away because of failed stabilization, and
then not being drawn or accounted for in any way. So now we still drop
them, but at least we account for them. Going to try to tweak some of
the tracking & stabilization parameters so we drop less, but some of the
drops are actually good and we probably never wanted those tracks
anyway.