Debugging: hand-engineered "discriminator"
Built a special "discriminator" that just takes some target video, gets
the image discriminator activations, and then just does L2 loss between
the target activations and the actual activations. Uses exactly the same
generator as the master branch. Trying to figure out if there is a
problem with the generator.